end to end test working with QR
