+++ packages/anon-aadhaar-react/src/components/ScanQR.tsx
@@ -13,27 +13,23 @@ export const ScanQR: FunctionComponent<ScanQRProps> = ({
 }) => {
   const [showScanner, setShowScanner] = useState(true)
 
-  return (
-    <>
-      {showScanner && (
-        <QrReader
-          containerStyle={{ width: '90%' }}
-          videoStyle={{ height: '70%' }}
-          // scanDelay={300}
-          constraints={{ facingMode: 'environment' }}
-          onResult={(result, error) => {
-            if (result?.getText()) {
-              setQrStatus(AadhaarQRValidation.QR_CODE_SCANNED)
-              setQrData(result.getText())
-              setShowScanner(false)
-            }
+  return showScanner ? (
+    <QrReader
+      containerStyle={{ width: '90%' }}
+      videoStyle={{ height: '70%' }}
+      // scanDelay={300}
+      constraints={{ facingMode: 'environment' }}
+      onResult={(result, error) => {
+        if (result?.getText()) {
+          setQrStatus(AadhaarQRValidation.QR_CODE_SCANNED)
+          setQrData(result.getText())
+          setShowScanner(false)
+        }
 
-            if (error) {
-              console.info(error)
-            }
-          }}
-        />
-      )}
-    </>
-  )
+        if (error) {
+          console.info(error)
+        }
+      }}
+    />
+  ) : null
 }
