+++ packages/anon-aadhaar-react/src/prove.ts
@@ -1,4 +1,4 @@
-import { SerializedPCD } from '@pcd/pcd-types'
+import { ArgumentTypeName, SerializedPCD } from '@pcd/pcd-types'
 import {
   prove,
   PCDInitArgs,
@@ -9,10 +9,13 @@ import {
   VK_URL,
   ZKEY_URL,
   WASM_URL,
+  extractWitness,
+  splitToWords,
 } from 'anon-aadhaar-pcd'
+import { fetchRawPublicKey } from './util'
 
 /**
- * proveWithWebProver is a function that generates proofs using the web-based proving system of anon-aadhaar.
+ * proveAndSerialize is a function that generates proofs using the web-based proving system of anon-aadhaar.
  * It takes an IdentityPCDArgs argument and returns a Promise containing the generated IdentityPCD and its
  * serialized form (SerializedPCD).
  *
@@ -42,4 +45,50 @@ export const proveAndSerialize = async (
   const serialized = await serialize(pcd)
 
   return { pcd, serialized }
+}
+
+/**
+ * processArgs is a function that generates the arguments to pass to the prover from the QR Data scanned.
+ *
+ * @param {string} qrData - The arguments required to generate the IdentityPCD.
+ * @returns A Promise containing the generated args object containing the parameters needed to generate a proof.
+ *
+ * @see {@link https://github.com/privacy-scaling-explorations/anon-aadhaar}
+ *   For more information about the underlying circuit and proving system.
+ */
+export const processArgs = async (
+  qrData: string,
+  testing: boolean,
+): Promise<AnonAadhaarPCDArgs> => {
+  const certificate = await fetchRawPublicKey(
+    `https://www.uidai.gov.in/images/authDoc/${
+      testing ? 'uidai_prod_cdup' : 'uidai_offline_publickey_26022021'
+    }.cer`,
+  )
+
+  if (!certificate) throw Error('Error while fetching public key.')
+
+  const witness = await extractWitness(qrData, certificate)
+
+  if (witness instanceof Error) throw new Error(witness.message)
+
+  const args: AnonAadhaarPCDArgs = {
+    padded_message: {
+      argumentType: ArgumentTypeName.StringArray,
+      value: witness.paddedMessage,
+    },
+    message_len: {
+      argumentType: ArgumentTypeName.Number,
+      value: witness.messageLength.toString(),
+    },
+    signature: {
+      argumentType: ArgumentTypeName.StringArray,
+      value: splitToWords(witness.signatureBigint, BigInt(64), BigInt(32)),
+    },
+    modulus: {
+      argumentType: ArgumentTypeName.StringArray,
+      value: splitToWords(witness.modulusBigint, BigInt(64), BigInt(32)),
+    },
+  }
+  return args
 }
