+++ packages/anon-aadhaar-react/src/verifySignature.ts
@@ -1,0 +1,67 @@
+import { convertBigIntToByteArray, decompressByteArray } from 'anon-aadhaar-pcd'
+import { fetchRawPublicKey, str2ab } from './util'
+import { pki } from 'node-forge'
+
+export const verifySignature = async (
+  n: string,
+  testing: boolean,
+): Promise<boolean> => {
+  console.log('is testing: ', testing)
+  const bigIntData = BigInt(n)
+
+  const byteArray = convertBigIntToByteArray(bigIntData)
+
+  const decompressedByteArray = decompressByteArray(byteArray)
+
+  // Read signature data
+  const signature = decompressedByteArray.slice(
+    decompressedByteArray.length - 256,
+    decompressedByteArray.length,
+  )
+
+  const signedData = decompressedByteArray.slice(
+    0,
+    decompressedByteArray.length - 256,
+  )
+
+  const certificate = await fetchRawPublicKey(
+    `https://www.uidai.gov.in/images/authDoc/${
+      testing ? 'uidai_prod_cdup' : 'uidai_offline_publickey_26022021'
+    }.cer`,
+  )
+
+  if (!certificate) throw Error('Error while fetching public key.')
+
+  const publicKey = pki.certificateFromPem(certificate).publicKey
+  const publicKeyPem = pki.publicKeyToPem(publicKey)
+
+  // fetch the part of the PEM string between header and footer
+  const pemHeader = '-----BEGIN PUBLIC KEY-----'
+  const pemFooter = '-----END PUBLIC KEY-----'
+  const pemContents = publicKeyPem.substring(
+    pemHeader.length,
+    publicKeyPem.length - pemFooter.length - 2,
+  )
+
+  // base64 decode the string to get the binary data
+  const binaryDerString = window.atob(pemContents)
+  // convert from a binary string to an ArrayBuffer
+  const binaryDer = str2ab(binaryDerString)
+
+  const pk = await window.crypto.subtle.importKey(
+    'spki',
+    binaryDer,
+    { name: 'RSASSA-PKCS1-v1_5', hash: 'SHA-256' },
+    true,
+    ['verify'],
+  )
+
+  const isSignatureValid = await window.crypto.subtle.verify(
+    { name: 'RSASSA-PKCS1-v1_5' },
+    pk,
+    signature.buffer,
+    signedData.buffer,
+  )
+
+  return isSignatureValid
+}
