+++ packages/anon-aadhaar-react/tsconfig.json
@@ -14,6 +14,6 @@
     "noUnusedParameters": true,
     "allowSyntheticDefaultImports": true
   },
-  "include": ["src/*", "../anon-aadhaar-pcd/src/extractWitness.ts"],
+  "include": ["src/*"],
   "exclude": ["mocha.cjs", "eslintrc.cjs"]
 }
